Do all routing in a single route for overflow.
For some unknown reason, the json module was messing up the JobRouter
environment.  This removes the use of JSON (converting everything
to classads), which seems to work.

NOTE NOTE: once merged, routing is a no-op until condor_job_router
is restarted (as the unified_utils module must be reloaded).

Because the route is completely applied in a single step, this is
much faster than the old overflow method.